refactor: Change package directories
