+++ modules/common/src/main/java/org/vstu/meaningtree/languages/LanguageTranslator.java
@@ -9,11 +9,11 @@ import org.vstu.meaningtree.MeaningTree;
 import org.vstu.meaningtree.exceptions.MeaningTreeException;
 import org.vstu.meaningtree.exceptions.UnsupportedConfigParameterException;
 import org.vstu.meaningtree.languages.configs.*;
-import org.vstu.meaningtree.languages.query.CompiledTSQuery;
-import org.vstu.meaningtree.languages.query.ParseSession;
-import org.vstu.meaningtree.languages.query.QueryResult;
-import org.vstu.meaningtree.languages.templates.TemplateEngine;
-import org.vstu.meaningtree.languages.templates.TemplateRepository;
+import org.vstu.meaningtree.languages.helpers.query.CompiledTSQuery;
+import org.vstu.meaningtree.languages.helpers.query.ParseSession;
+import org.vstu.meaningtree.languages.helpers.query.QueryResult;
+import org.vstu.meaningtree.languages.helpers.templates.TemplateEngine;
+import org.vstu.meaningtree.languages.helpers.templates.TemplateRepository;
 import org.vstu.meaningtree.nodes.Node;
 import org.vstu.meaningtree.utils.Experimental;
 import org.vstu.meaningtree.utils.Label;
