+++ modules/common/src/main/java/org/vstu/meaningtree/languages/helpers/query/ByteRange.java
@@ -1,4 +1,4 @@
-package org.vstu.meaningtree.languages.query;
+package org.vstu.meaningtree.languages.helpers.query;
 
 public record ByteRange(int startByte, int endByte) {
     public ByteRange {
