+++ modules/common/src/main/java/org/vstu/meaningtree/languages/helpers/query/QueryResult.java
@@ -1,4 +1,4 @@
-package org.vstu.meaningtree.languages.query;
+package org.vstu.meaningtree.languages.helpers.query;
 
 import org.treesitter.TSNode;
 import org.vstu.meaningtree.nodes.Node;
