+++ modules/common/src/main/java/org/vstu/meaningtree/languages/helpers/templates/TemplateEngine.java
@@ -1,4 +1,4 @@
-package org.vstu.meaningtree.languages.templates;
+package org.vstu.meaningtree.languages.helpers.templates;
 
 import java.util.Map;
