+++ modules/common/src/main/java/org/vstu/meaningtree/languages/helpers/templates/TemplateRepository.java
@@ -1,4 +1,4 @@
-package org.vstu.meaningtree.languages.templates;
+package org.vstu.meaningtree.languages.helpers.templates;
 
 public interface TemplateRepository {
     String getTemplateSource(String templateKey);
